+++ commons/src/test/java/com/stratio/morphlines/commons/BaseTest.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright (C) 2015 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.stratio.morphlines.commons;
+
+import com.google.common.base.Preconditions;
+import com.typesafe.config.Config;
+import org.apache.commons.io.IOUtils;
+import org.kitesdk.morphline.api.Record;
+import org.kitesdk.morphline.base.Compiler;
+
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+
+import static org.junit.Assert.assertEquals;
+
+/**
+ * Created by dvazquez on 28/11/16.
+ */
+public class BaseTest {
+
+    protected Config parse(String file, Config... overrides) throws IOException {
+        File tmpFile = File.createTempFile("morphlines_", ".conf");
+        IOUtils.copy(getClass().getResourceAsStream(file), new FileOutputStream(tmpFile));
+        Config config = new Compiler().parse(tmpFile, overrides);
+        config = config.getConfigList("morphlines").get(0);
+        Preconditions.checkNotNull(config);
+        return config;
+    }
+
+    protected void checkFieldValue(Record record, String fieldName, String expectedValue){
+        assertEquals(1, record.get(fieldName).size());
+        assertEquals(expectedValue, record.get(fieldName).get(0));
+    }
+
+}