+++ commons/src/test/java/com/stratio/morphlines/commons/FieldFilterTest.java
@@ -15,11 +15,8 @@
  */
 package com.stratio.morphlines.commons;
 
-import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
-import com.typesafe.config.Config;
-import org.apache.commons.io.IOUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -27,29 +24,17 @@ import org.junit.runners.JUnit4;
 import org.kitesdk.morphline.api.Command;
 import org.kitesdk.morphline.api.MorphlineContext;
 import org.kitesdk.morphline.api.Record;
-import org.kitesdk.morphline.base.Compiler;
 
-import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.List;
 
 import static org.fest.assertions.Assertions.assertThat;
 
 @RunWith(JUnit4.class)
-public class FieldFilterTest {
+public class FieldFilterTest extends BaseTest {
 
     @Before
     public void setUp() {
-    }
-
-    protected Config parse(String file, Config... overrides) throws IOException {
-        File tmpFile = File.createTempFile("morphlines_", ".conf");
-        IOUtils.copy(getClass().getResourceAsStream(file), new FileOutputStream(tmpFile));
-        Config config = new Compiler().parse(tmpFile, overrides);
-        config = config.getConfigList("morphlines").get(0);
-        Preconditions.checkNotNull(config);
-        return config;
     }
 
     private Record record(String...args) {