+++ commons/src/test/java/com/stratio/morphlines/commons/TimeFilterTest.java
@@ -36,7 +36,7 @@ import com.google.common.base.Preconditions;
 import com.typesafe.config.Config;
 
 @RunWith(JUnit4.class)
-public class TimeFilterTest {
+public class TimeFilterTest extends BaseTest{
 
     private static final String MORPH_CONF_FILE = "/timefilter/timeFilter.conf";
 
@@ -82,13 +82,5 @@ public class TimeFilterTest {
         assertThat(records.size()).isEqualTo(0);
     }
 
-    protected Config parse(String file, Config... overrides) throws IOException {
-        File tmpFile = File.createTempFile("morphlines_", ".conf");
-        IOUtils.copy(getClass().getResourceAsStream(file), new FileOutputStream(tmpFile));
-        Config config = new Compiler().parse(tmpFile, overrides);
-        config = config.getConfigList("morphlines").get(0);
-        Preconditions.checkNotNull(config);
-        return config;
-    }
 
 }