+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 ## 0.2.0 (upcoming)
 
 * Added coveralls badge
+* Added new command ```vgrok```
 
 ## 0.1.0 (September 2015)